feat: replace ChartScroll with ChartHome component for enhanced dashboard functionality
